chore: move `RemapNamespaceAttribute` to `api`
