+++ src/main/kotlin/earth/terrarium/cloche/Attributes.kt
@@ -82,13 +82,3 @@ internal object ModTransformationStateAttribute {
     fun of(target: MinecraftTarget, compilation: TargetCompilation, state: String) =
         lowerCamelCaseName(target.featureName, compilation.featureName, state)
 }
-
-object RemapNamespaceAttribute {
-    @JvmField
-    val ATTRIBUTE = Attribute.of("earth.terrarium.cloche.remapNamespace", String::class.java)
-
-    const val INITIAL = "none"
-    const val OBF = "obf"
-    const val SEARGE = "srg"
-    const val INTERMEDIARY = "intermediary"
-}
